Dhan-Shomadhan/requirements.txt
# Deep learning frameworks
--index-url https://download.pytorch.org/whl/cu121
torch>=2.5.1
torchvision>=0.15.0

# YOLO
ultralytics>=8.0.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Image processing and visualization
opencv-python>=4.5.0
Pillow>=9.0.0
matplotlib>=3.5.0

# Utilities
tqdm>=4.60.0
